web: refactor speciescorrect into components
